refactor unsafe code out again
