+++ statime-linux/src/network/mod.rs
@@ -2,3 +2,5 @@ mod control_message;
 mod interface;
 pub mod linux;
 pub mod linux_syscall;
+mod raw_udp_socket;
+mod timestamped_udp_socket;
